requirements/dev.txt
arcgis-sdk>=0.0.9
django-core-flavor>=0.0.14
django-filter>=1.0.2
django-oauth-toolkit>=0.12.0
django-orders-flavor>=0.0.2
django-polymorphic>=1.2
django-taggit>=0.22.1
djangorestframework-filters>=0.10.0
djangorestframework==3.6.3
Pillow>=4.1.1
psycopg2>=2.6.2
social-auth-app-django>=1.2.0
sorl-thumbnail>=12.3
